+++ R/internal.r
@@ -3,13 +3,11 @@
 #' Output index and name for features that will be ignored
 #' @param dt input data object.
 #' @param maxcat maximum categories allowed for each discrete feature.
-#' @return a named vector containing indices of features to be ignored.
+#' @return a named vector containing number of categories for to-be-ignored features.
 #' @import data.table
 .ignoreCat <- function(dt, maxcat) {
 	if (!is.data.table(dt)) dt <- data.table(dt)
-	n_cat <- sapply(dt, function(x) {
-		length(unique(x))
-	})
+	n_cat <- vapply(dt, function(x) {length(unique(x))}, 0)
 	n_cat[which(n_cat > maxcat)]
 }