+++ R/introduce.r
@@ -3,7 +3,7 @@
 #' Describe basic information for input data.
 #' @param data input data
 #' @keywords introduce
-#' @return Describe basic information in a \link{data.frame}:
+#' @return Describe basic information in input data class:
 #' \itemize{
 #'   \item{rows: number of rows}
 #'   \item{columns: number of columns}
@@ -21,15 +21,25 @@
 #' introduce(mtcars)
 
 introduce <- function(data) {
-  split_data <- split_columns(data)
-  data.frame(
-    "rows" = nrow(data),
-    "columns" = ncol(data),
-    "discrete_columns" = split_data[["num_discrete"]],
-    "continuous_columns" = split_data[["num_continuous"]],
-    "all_missing_columns" = split_data[["num_all_missing"]],
-    "total_missing_values" = sum(is.na(data)),
-    "total_observations" = nrow(data) * ncol(data),
-    "memory_usage" = as.numeric(object.size(data))
-  )
+	## Check and set to data.table
+	is_data_table <- is.data.table(data)
+	data_class <- class(data)
+	if (!is.data.table(data)) data <- data.table(data)
+
+	split_data <- split_columns(data)
+
+	output <- data.table(
+		"rows" = nrow(data),
+		"columns" = ncol(data),
+		"discrete_columns" = split_data[["num_discrete"]],
+		"continuous_columns" = split_data[["num_continuous"]],
+		"all_missing_columns" = split_data[["num_all_missing"]],
+		"total_missing_values" = sum(is.na(data)),
+		"total_observations" = nrow(data) * ncol(data),
+		"memory_usage" = as.numeric(object.size(data))
+	)
+
+	## Set data class back to original
+	if (!is_data_table) class(output) <- data_class
+	output
 }